requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
defusedxml==0.6.0
Django==3.1
django-allauth==0.42.0
django-bootstrap-modal-forms==2.0.0
django-filter==2.4.0
django-phonenumber-field==5.0.0
django-stdimage==5.1.1
idna==2.10
oauthlib==3.1.0
phonenumbers==8.12.9
Pillow==7.2.0
progressbar2==3.51.4
psycopg2==2.8.5
pycparser==2.20
python-utils==2.4.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
